Move type annotations into types.ts
